requirements.txt
torch
torchvision
numpy
matplotlib
opencv-python-headless
hydra-core
Pillow
